fix(tokenizer): improve remote tokenizer reliability and performance
- Fix exponential backoff algorithm to avoid float type error in bit shift
- Add proper resource cleanup with defer for HTTP response bodies
- Implement concurrency control with semaphore (default: 1000 for gateway)
- Increase connection pool limits for cluster gateway usage
  - MaxIdleConns: 100 -> 2000
  - MaxIdleConnsPerHost: 10 -> 100
  - IdleConnTimeout: 90s -> 10s for faster cleanup
- Enhance URL validation to check for missing scheme
- Fix type conversion safety in utils with int32 handling
- Add comprehensive test coverage for HTTP client and remote tokenizer
- Remove /v1 prefix from test paths and sglang-related tests
- Update error messages for consistency

These changes make the tokenizer more suitable for high-throughput
cluster gateway scenarios while preventing resource exhaustion.

+++ pkg/utils/tokenizer/utils.go
@@ -21,11 +21,14 @@ import (
 )
 
 // intToByteArray converts int array to byte array in BigEndian format
+// It uses int32 to preserve sign and validate token values
 func intToByteArray(intArray []int) []byte {
 	// Pre-allocate buffer for better performance
 	buf := make([]byte, len(intArray)*4)
 	for i, num := range intArray {
-		binary.BigEndian.PutUint32(buf[i*4:], uint32(num))
+		// Use int32 to preserve sign and ensure valid range
+		// Token IDs should typically be positive, but we handle negative values safely
+		binary.BigEndian.PutUint32(buf[i*4:], uint32(int32(num)))
 	}
 	return buf
 }
